Pantalla ConfirmaCompra
Se da funcionalidad para que muestre dgv con productos comprados.

+++ Dominio/ItemCarrito.cs
@@ -7,7 +7,7 @@ namespace Dominio
 {
     public class ItemCarrito
     {
-        public int IDItem { get; set; }
+        public short IDItem { get; set; }
         public string NombreItem { get; set; }
         public int Cantidad { get; set; }
         public decimal Precio { get; set; }
